Update patch set 2
Patch Set 2:

I installed it using docker-compose up and it actually worked everything and started two containers, one with gerrit on port 18080 and one postgres. I then was able to create a project and submit a change to gerrit via web. Unsure however where and how to check if it is currently working. Probably it would be important if there is the README.md a simple IVP procedure (installation and verification procedure) that can help in understanding if the container is really working and exactly what to expect. Unless I have missed this piece of information.

Patch-set: 2
Reviewer: Claudio Pacchiega <1009089@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0